// src/enums.ts
// Our own enum
enum FINGER_ENUM {
    thumb = "thumb",
    indexFinger = "indexFinger",
    middleFinger = "middleFinger",
    ringFinger = "ringFinger",
    pinky = "pinky",
}
enum GESTURES_ENUM {
    play = "play",
    pause = "pause",
    volumeUp = "volume-up",
    volumeDown = "volume-down",
    forward = "forward",
    backward = "backward"
}

export { FINGER_ENUM, GESTURES_ENUM }